+++ web-frontend/modules/core/plugins/global.js
@@ -13,6 +13,7 @@ import SwitchInput from '@baserow/modules/core/components/SwitchInput'
 import Copied from '@baserow/modules/core/components/Copied'
 import MarkdownIt from '@baserow/modules/core/components/MarkdownIt'
 import DownloadLink from '@baserow/modules/core/components/DownloadLink'
+import FormElement from '@baserow/modules/core/components/FormElement'
 
 import lowercase from '@baserow/modules/core/filters/lowercase'
 import uppercase from '@baserow/modules/core/filters/uppercase'
@@ -39,6 +40,7 @@ Vue.component('SwitchInput', SwitchInput)
 Vue.component('Copied', Copied)
 Vue.component('MarkdownIt', MarkdownIt)
 Vue.component('DownloadLink', DownloadLink)
+Vue.component('FormElement', FormElement)
 
 Vue.filter('lowercase', lowercase)
 Vue.filter('uppercase', uppercase)